+++ src/bindings/scripts/scripts/zone/tempest_keep/botanica/boss_warp_splinter.cpp
@@ -151,7 +151,7 @@ struct TRINITY_DLL_DECL boss_warp_splinterAI : public ScriptedAI
 
     void SummonTreants()
     {
-        for(int i = 0; i < 6; ++i)
+        for(uint8 i = 0; i < 6; ++i)
         {
             float angle = (M_PI / 3) * i;
